Fixed most Newport issues. Update low level vi's
Still need to fix way to read position from Newport motors. The proper
way to do it per Newport is timing out possibly from the use of older
VISA calls. I can use a RAW call to 01TPnn? to poll it and it works of
it is a matter of debugging their vi